requirements.txt
colorama~=0.4.5
requests~=2.28.1
beautifulsoup4~=4.11.1
setuptools~=65.3.0
Pillow~=9.2.0